Relocate roosa command
